rename repository gogen -> egoist
