+++ egoist/_fnspec.py
