+++ egoist/generate.py
