+++ egoist/langhelpers.py
