+++ egoist/naming.py
