+++ egoist/prestringutil.py
