+++ egoist/scan.py
