+++ egoist/types.py
