+++ examples/e2e/generate/00simple/definitions.py
@@ -1,9 +1,9 @@
-from gogen import runtime
+from egoist import runtime
 
 
 def hello(*, name: str) -> None:
     """hello message"""
-    from gogen.generate import cli
+    from egoist.generate import cli
 
     with runtime.generate(cli):
         runtime.printf("hello %s\n", name)
@@ -11,7 +11,7 @@ def hello(*, name: str) -> None:
 
 def byebye(*, name: str) -> None:
     """byebye message"""
-    from gogen.generate import cli
+    from egoist.generate import cli
 
     with runtime.generate(cli):
         runtime.printf("byebye %s\n", name)
