+++ examples/e2e/init/00hello/definitions.py
@@ -1,9 +1,9 @@
-from gogen import runtime
+from egoist import runtime
 
 
 def hello(*, name: str) -> None:
     """hello message"""
-    from gogen.generate import cli
+    from egoist.generate import cli
 
     with runtime.generate(cli):
         runtime.printf("hello %s\n", name)
