+++ setup.cfg
@@ -1,13 +1,13 @@
 # see: https://setuptools.readthedocs.io/en/latest/setuptools.html#specifying-values
 [metadata]
-name = gogen
+name = egoist
 version = file:VERSION
 
 summary = -
 long_description = file:README.md
 long_description_content_type = text/markdown; charset=UTF-8; variant=GFM
 
-url = https://github.com/podhmo/gogen
+url = https://github.com/podhmo/egoist
 author = podhmo
 author_email = ababjam61+github@gmail.com
